+++ NEWS.md
@@ -13,6 +13,7 @@ This may have been addressed at 64a308ba50b9119108e8946737460de5997b805b by addi
 
 # vcfR 1.9.0
 Released on CRAN 201X-XX-XX
+* `extract.haps()` reports the correct number of variants processed when verbose.
 * The square brackets ([]) handle @gt slots with no samples.
 * `vcfR2loci()` now has the option `return.alleles = FALSE`.
 * `vcfR2genind()` now has the option `retrun.alleles = FALSE`.